Fix transaction propagation and refactor
