+++ src/test/java/uk.gov.ons.census.action/messaging/ConsumeAndPublishIT.java
@@ -4,12 +4,9 @@ import static junit.framework.TestCase.assertNull;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.Assert.assertNotNull;
 
-import java.io.IOException;
 import java.io.StringReader;
 import java.time.OffsetDateTime;
 import java.util.HashMap;
-import java.util.List;
-import java.util.Random;
 import java.util.UUID;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.TimeUnit;
@@ -35,8 +32,6 @@ import uk.gov.ons.census.action.model.dto.instruction.ActionInstruction;
 import uk.gov.ons.census.action.model.entity.ActionPlan;
 import uk.gov.ons.census.action.model.entity.ActionRule;
 import uk.gov.ons.census.action.model.entity.ActionType;
-import uk.gov.ons.census.action.model.entity.Case;
-import uk.gov.ons.census.action.model.entity.UacQidLink;
 import uk.gov.ons.census.action.model.repository.ActionPlanRepository;
 import uk.gov.ons.census.action.model.repository.ActionRuleRepository;
 import uk.gov.ons.census.action.model.repository.CaseRepository;
@@ -75,8 +70,7 @@ public class ConsumeAndPublishIT {
   }
 
   @Test
-  public void checkRecievedEventsAreEmitted()
-      throws IOException, InterruptedException, JAXBException {
+  public void checkRecievedEventsAreEmitted() throws InterruptedException, JAXBException {
     // Given
     BlockingQueue<String> outputQueue = rabbitQueueHelper.listen(outboundPrinterQueue);
 
@@ -112,7 +106,7 @@ public class ConsumeAndPublishIT {
 
   @Test
   public void checkCaseWithNoLinkedUACQuidDoesntSendThenWorksWhenUACAdded()
-      throws IOException, InterruptedException, JAXBException {
+      throws InterruptedException, JAXBException {
     // Given
     BlockingQueue<String> outputQueue = rabbitQueueHelper.listen(outboundPrinterQueue);
 
@@ -157,7 +151,6 @@ public class ConsumeAndPublishIT {
   }
 
   private ActionPlan setUpActionPlan(String name, String desc) {
-    Random random = new Random();
     ActionPlan actionPlan = new ActionPlan();
     actionPlan.setName(name);
     actionPlan.setDescription(desc);
@@ -168,11 +161,11 @@ public class ConsumeAndPublishIT {
   private void checkExpectedMessageNotReceived(BlockingQueue<String> queue)
       throws InterruptedException {
     String actualMessage = queue.poll(10, TimeUnit.SECONDS);
-    assertNull("Recieved Message, expectedNone", actualMessage);
+    assertNull("Recieved Message, expected none", actualMessage);
   }
 
   private ActionInstruction checkExpectedMessageReceived(BlockingQueue<String> queue)
-      throws IOException, InterruptedException, JAXBException {
+      throws InterruptedException, JAXBException {
     String actualMessage = queue.poll(10, TimeUnit.SECONDS);
     assertNotNull("Did not receive message before timeout", actualMessage);
 
@@ -209,16 +202,5 @@ public class ConsumeAndPublishIT {
     actionRule.setActionPlan(actionPlan);
 
     return actionRule;
-  }
-
-  private void setUpQidLinksForCases(List<Case> cases) {
-    cases.forEach(
-        caze -> {
-          String uac = caze.getCaseId().toString() + "uac";
-          UacQidLink uacQidLink = new UacQidLink();
-          uacQidLink.setUac(uac);
-
-          uacQidLinkRepository.save(uacQidLink);
-        });
   }
 }
